fix: reorganise, types
